Update patch set 4
Patch Set 4:

> Patch Set 4:
>
> Flashing an individual physical partition when half upgraded seems like a bit of a corner case.  Maybe worth showing a warning, but maybe okay to allow too; since that won't fundamentally cause any issues and we don't have the ability to upgrade the rest of the physical partitions to logical if we're only flashing a single one.

Sounds good.

> As a side note, is there anything that we're doing to prevent users from flashing a physical partition on retrofit devices?  Aka prevent flashing the physical system_a, which is actually a super partition?  We'd need to block this in bootloader fastboot especially too.

Blocking in the bootloader is hard, since we need to keep it working for flashing in factory. But maybe we could block it in host fastboot. We'd have to parse super_empty.img to do that, which is blocked on win32 support for c++17 - unless we just make the liblp linkage linux/mac-only for now.

Patch-set: 4
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0