Update patch set 1
Patch Set 1:

(2 comments)

I understand the need for something like this, but it's a bit scary to me.  Especially since with first stage mount, we've been moving to a world where we're starting more and more services before /data is mounted, but now we're going to try to revert back to the old way.  It's definitely going to increase boot time, and I worry that some services may not tolerate their dependencies starting so much later as well as we'd like.

Patch-set: 1
Reviewer: Gerrit User 1105075 <1105075@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>