Update patch set 3
Patch Set 3:

(2 comments)

normally i'd grumble about adding this to system/core/, but in a sense this is part of init (but can't be as long as init is a static binary).

one higher-level question: is boringssl going to be updated via the APEX stuff? will this actually get the right .so if so?

Patch-set: 3