Update patch set 5
Patch Set 5:

> normally i'd grumble about adding this to system/core/, but in a sense this is part of init (but can't be as long as init is a static binary).

Would we want this to be a part of init directly?  Init (2nd stage) isn't static, so this could be done.  That probably is a good idea here afterall, given how trivial this program is otherwise.  Something like update_verifier makes sense to be separate, but this is simple enough that it might be worth just including in.

Patch-set: 5
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>