Update patch set 5
Patch Set 5:

This scares me a little.  I guess there's no other way around this if the newly flashed partition has increased in size, but essentially that means if I'm using scratch for changes on system and I flash a new vendor, that I lose those changes to system, right?

I'd almost think that we'd maybe want this to be a behavior where users need to 'force' this or 'wipe-scratch' or something?

Patch-set: 5
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>