Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> This scares me a little.  I guess there's no other way around this if the newly flashed partition has increased in size, but essentially that means if I'm using scratch for changes on system and I flash a new vendor, that I lose those changes to system, right?
> 
> I'd almost think that we'd maybe want this to be a behavior where users need to 'force' this or 'wipe-scratch' or something?

I would have no problems with any alternatives if it were not for the fact that fastboot is likely to be performed on a bricked (can not boot into adb) device ...

[TL;DR]

The alternative would be to report back from fastboot a message that suggests that the flash failed because scratch was present, and ask the user to do an adb enable-verity to wipe out the overrides. Problem is, they are in fastboot and likely can't reboot successfully to get to adb to do an enable-verity to wipe it. Catch 22. So I _need_ this to ensure that "scratch" never prevents a flash because it is too late to fix it ...

The problem I have with that is OTAs do _not_ have this option (to report back mitigation responses like this) and needs to work _always_. Since it is changing everything, "scratch" does not matter.

However, if they (fastboot and OTA) call fs_mgr_overlayfs_teardown() before they start applying the updates to all the filesystems, then these problems goes away. It then only remains for fastboot flash <individual partition> ...

If everyone (including custom partner code that performs OTAs) calls fs_mgr_overlayfs_teardown() when they intend to change _everything_ (fastboot flashall and OTA), this code will never be reached and it merely turns into insurance to ensure that we _never_ fail a flash b/c of scratch and grant ourselves the (caveat and) privilege to make scratch fragile.

Patch-set: 5
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>