Update patch set 7
Patch Set 7: Verified+1

> Patch Set 7: Code-Review+2
> 
> Okay, thanks for the explanation.
> 
> One last question: I can't find the patch anymore (I think it was a previous patchset for this commit actually), but there was code that would delete scratch in the case that ResizePartition() fails.  Is that still the plan?

You were nervous about deleting scratch to permit ResizePartition, so I removed that part of this CL as it is controversial (albeit simple). I accepted that if _everyone_ does their job (fs_mgr_overlayfs_teardown()) in preparation for flashing (everything), then it will be a minor corner case, especially with _this_ margin in place. I will continue to evaluate the need to (automagically) remove scratch in ResizePartition and if it appears to be required in this form or another, will add it later.

Patch-set: 7
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1