Update patch set 10
Patch Set 10:

> Oh, by the way, are you sure that this works? (if you manually remove bionic libs from system/lib, did it work? It didn't work for me...)

No. If it was simple to try out I'd do it, but I need to learn how to (re)mount / rw first, or else fix the packaging to remove those libs in the build, but I gather that needs support for stub libs in the cc_library rules?

In any case, I'm not aiming to actually remove those libs in this CL, and since it doesn't regress as long as all libs are duplicated I thought we'd address it separately.

Patch-set: 10
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>