Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > I still don't understand why this isn't refactored into a function that checks the verity state for a specific fstab_entry and then fs_mgr_update_verity_state() calls that function in a loop for the original callers.  Or, since there's only one other caller, just refactoring the whole thing and moving the looping logic to the caller, is probably better; it'll allow removing the callback too, since the function that takes a single fstab_entry can return everything directly.
> > 
> > fs_mgr_overlayfs.cpp, which is where I'm assuming the problems are stemming from, only uses this function to check for each fstab_entry that it checks if it has verity enabled already.  It already would be better off with this single function and the error reporting could be much more direct.
> 
> overlayfs collects them all (an early version checked for each entry, but that created even more noise and was the real reason we got all up in arms over the logspam), prior to looping through candidates so there is no reason for a refactor. AFAIC the logspam problem was solved when we collected them all just once.

I think fundamentally it'd still be a better idea to refactor this.  There's nothing fundamentally dependent between each of these entries that it would make sense to collect them up front.  It makes the code slightly less clear too, since instead of checking "is verity enabled for X" it needs to check "does X exist in this list".

> 
> The problem _today_ is that this error message creates FUD and engineering churn. The _best_ answer may be to not print anything at all ... except that someone may wish to know that (any) verity information is missing. So re-opening this CL is meant to discuss an appropriate solution so that we do not waste eng-man-days over the misinterpretation of this messaging.

I think if we refactor it's easier to change how these errors are handled.  Or even just drop them entirely.  I agree with you that they're not particularly useful anyway.  Maybe I'll just push the CL that I'm thinking of and we can discuss that?  I think regardless of this issue, refactoring that function is a good idea.

Patch-set: 8
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>