Update patch set 6
Patch Set 6: Presubmit-Ready+1

(1 comment)

The remaining issue is a VNDK compliance issue.  It looks like a bug in the VNDK header checker though, since the only difference in that with this change it's referencing the full path including my workspace directory and that's causing the issue.  I filed b/122736285 for this.

I'm looking at ways to work around it as well in the meantime.

Patch-set: 6
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1