Update patch set 7
Patch Set 7:

As was reported at b/120266448#comment38 this fix is incomplete. It does not work when pushing to an opened executable file (like linker/linker64).

This is happening because the files which are mount points cannot be unlinked; 

Root cause of this problem is on the design that we are bind-mounting /system/lib/libc.so into /system/lib/libc.so. (and then /apex/.../libc.so into the same mount point). Since the path /system/lib/libc.so is both a mount source and a mount point, unlinking the path always fails with -EBUSY. Since the file isn't unlinked, opening the file for writing fails.

Proper fix would be re-locating the real libc.so to under/system/lib/bootstrap and bind mount it to /system/lib/libc.so. Then unlinking /system/lib/bootstrap/libc.so would be successful as it is no longer a mount point. I've locally (with shell commands) confirmed that this fixes the problem.

I am working on to make the change, but the extent is quite big: from build system, bionic, sepolicy, to init. This will take some time. Sorry. :(

Patch-set: 7
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0