Update patch set 1
Patch Set 1:

> Somewhere around here seems like the right place to me: https://android.googlesource.com/platform/system/core/+/master/adb/adb.cpp#114
> 
> Also, now that I think about it, it would be pretty bad if the system server deadlocked or something, and we ended up blocking trying to write to it in adbd, which means we probably shouldn't actually use WriteFdExactly? The best solution I can think of is to make the socket a SOCK_SEQPACKET socket and switch the framework fd to nonblocking.

I was reviewing the logic again and ran some tests and while the function is named usb_disconnected it actually is the callback for both a USB and TCP disconnect. This callback is invoked if the key is not found in adb_keys and adbd_auth_confirm_key is called to confirm the key in the framework. Since this is working as intended from the transport's RunDisconnects function for both TCP and USB disconnects I don't think we want to move this, but the name of the function should be changed to something more appropriate to indicate it covers both types of disconnects. What do you think?

Patch-set: 1
Reviewer: Gerrit User 1294681 <1294681@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0