Update patch set 2
Patch Set 2:

What's the verdict on this?  From the bug it seemed like we weren't going to go forward with it, which is why I'm holding off on submitting it.  I'd still want to expand this to reboot too since it's only covering fsck right now.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>