Update patch set 9
Patch Set 9:

> Patch Set 9:
> 
> > The inconsistent content problem that you just mentioned is actually due to the kernel bug in sdcardfs (b/122428178). Because of the bug, we are currently *not* using multiple namespaces across post and pre-apexd processes. As a result, the libc from runtime APEX which is not for adbd and its children are visible to adbd. This will be fixed when we correctly have multiple namespaces. But given the tight time constraint, I am very skeptical that we can make it on time for Q.
> 
> It's unclear to me whether there's even a correct answer here, though. Some people will want to see the same libc.so that adbd uses, but stuff like using gdbserver over adb to debug an app that's using the libc.so in the apex will break, won't it?
> 

Oh, gdbserver is a good counter example. Yes, your are right. Well, I really don't have a perfect solution here too.

> (Do we really need to mess with mount namespaces instead of having everything just be in a different directory?)

That was the original design, but we soon found that there are many things that explicitly or implicitly depend on the fact that libc (and its friends including the dynamic linker) are at /system/lib and are loaded from the default "linker" namespace. Believe of not, this design with mount namespace is to minimize the impact of moving bionic to the APEX.

Patch-set: 9
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0