Update patch set 5
Patch Set 5:

NB: nnk@ there is nothing in this on-device command that can not be done manhandled by an exploit that takes advantage of the existence of su. Please evaluate the security implications of adding this command to userdebug devices that have verity disabled (given that all googlers, including executives, may have userdebug builds on their personal devices). In effect for devices that have r/w partitions, this is the _same_ as mount -o rw,remount. Devices that have erofs, squashfs, ext4 dedupe, right-sized partitions or Dynamic Partitions, it transparently deals with all the overlayfs mechanics. I can not say the same for the CLs that depend on this, they need a separate security review.

Patch-set: 5
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>