Update patch set 1
Patch Set 1:

(This CL should have been created as a draft for discussion)

runcon is not currently granted perms for sh or setprop, but it does not mean this CL is likely fragile security.

Purpose of adding this security is that an executable could decide that it can only run if service.adb.root is a value of 1; thus guaranteeing that the command is executed when USB has physical access.

Other techniques being considered would be confirming that the pdi parentage leads to a _real_ /system/bin/adbd; thus preventing random applications from using su to run the remount executable.

Patch-set: 1
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0