Update patch set 2
Patch Set 2:

> /.../ Martin, can you file a bug against the owner of the tests?

I filed b/123864775 as I mentioned above, but it's perhaps not against the owner(s) of the tests. I figured even enumerating all the affected tests and finding their owners is not trivial, or is it? (I wouldn't assume it's restricted to the tests that failed in b/123700170.)

Patch-set: 2
Reviewer: Gerrit User 1373864 <1373864@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0