Update patch set 4
Patch Set 4:

Was it this change that requires the new SEPolicy?  That's a bit strange if so, since it *should* be a no-op for Marlin.  Instead of faking the /system entry, it should just provide it with a real '/' entry during first stage mount.

Did you test with both this and the next change?  Do you want me to wait for the SELinux change before I submit this?

I'm going to give one last run of adb-remount-test.sh on blueline too.

Patch-set: 4
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>