Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> _why_ do we even have to add anything? can't you just have the build
> system automatically implement your advice of "When in doubt, put
> gtests in device-tests and APK tests in general-tests", and only
> require boilerplate if that's _not_ the case? (which would be when?
> the docs don't make it clear. do you actually mean "anything with
> native code should be device-tests, and anything Java-only should be
> general-tests"? so cc_test should automatically be in device-tests,
> and we should do this in the build system?)

Hey Elliot,

So a quick overview. First off we need to package the tests into a few known artifacts so we can find them for the execution. Modifying the test's build script seemed better than using a list like https://cs.corp.google.com/android/platform_testing/build/tasks/tests/native_test_list.mk?sq=package:%5Eandroid$&dr=C&g=0

As for the warning, I believe we only have Pass/Fail from TreeHugger, I'll double-check if that's true and if not add a warning. Running Test Mapping with Atest locally will warn the user if the test is not in the correct suites so that should alert them (if they run it locally). Why not fail if the test is not included? Well not all tests build for all targets so this seemed like the safer course.

Auto-including cc_tests into the suite would balloon the size of the tests zips and cause slowdown for all of Test Mapping. That said we are migrating to a model (designing phase now) where each test will build it's own zip. I am okay with auto-including cc_tests when we migrate to that model, but I included Dan Willemsen as the build team may not be onboard for that increase in capacity :/....

Patch-set: 1
Reviewer: Gerrit User 1088076 <1088076@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Gerrit User 1083173 <1083173@85c56323-6fa9-3386-8a01-6480fb634889>