Update patch set 12
Patch Set 12:

> Patch Set 12:
> 
> (3 comments)
> 
> Looks mostly good to me now.  I'm still interested in jmgao@'s comments on if this is sufficient to replace the code for `adb remount` as keeping those two use cases in line would be a good idea.

I think so? There's an unmount of /bionic (that seems like it really should be handled somewhere else) in remount_service that isn't in here, and there's also the question of whether this should be rebooting or if adb remount should be doing it, but it looks like all of the core functionality can be delegated to an exec of this binary. We should definitely switch remount_service to exec this, if only to see what's missing.

Patch-set: 12