Update patch set 12
Patch Set 12:

> Patch Set 12:
> 
> > Patch Set 12:
> > 
> > (3 comments)
> > 
> > Looks mostly good to me now.  I'm still interested in jmgao@'s comments on if this is sufficient to replace the code for `adb remount` as keeping those two use cases in line would be a good idea.
> 
> I think so? There's an unmount of /bionic (that seems like it really should be handled somewhere else) in remount_service that isn't in here, and there's also the question of whether this should be rebooting or if adb remount should be doing it, but it looks like all of the core functionality can be delegated to an exec of this binary. We should definitely switch remount_service to exec this, if only to see what's missing.

Let me percolate on this a bit longer (respin and test), good catch, I should pull in the /bionic adjustment as it is needed _today_.

I have not got a CL to actually switch to this command from adb, so I should add one so that I can start baking and testing it. We do have adb-remount-test that flexes a lot of things adb remount. I was more than willing to punt the effort of exec remount after Q, but hey, why not pull it in now that the bug that triggered this has a (different and temporary) workaround. We will support -R here (see later CL), and will evaluate the side-effects on adb operations shifting it to the remount exec of course.

Patch-set: 12
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>