Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Isn't any globally-writable directory a side channel? You can test for file presence with O_EXCL even if you can't open any files or list the directory. Am I missing something?

The directory wouldn't be writable on user builds. So the concern you raise isn't relevant here.

But yes, in the general case, a side channel leak exists whenever you have search permission in the directory ("x" bit in chmod, or "search" permission in SELinux), and the filenames in the directory are predictable. The second part is important - with unpredictible filenames you can grant search permissions but still not leak data. See, for example, the directories in /data/app. In that directory, we append a random string, so that can't be used as a side channel for installed apps.

Patch-set: 2
Reviewer: Gerrit User 1003966 <1003966@85c56323-6fa9-3386-8a01-6480fb634889>