Update patch set 2
Abandoned

Abandoning this change in favor of https://android-review.googlesource.com/c/platform/system/core/+/940866 , which is taking a broader approach towards solving this problem. In particular, in that change, the third argument is passed as a 32 bit value, instead of a size_t. Please see comments in that change.

I assume that we'll resolve these issues in that change, and this change isn't needed.

Patch-set: 2
Status: abandoned
Tag: autogenerated:gerrit:abandon