Update patch set 8
Patch Set 8: Code-Review-1

I am not sure we need to make such a huge surgery to first stage init for this.

All we want is modprobe / insmod correct? Why can't  we keep this very simple instead of bringing in all the complexities of .rc files etc.

Lets discuss this offline (in a doc may be?) before we go further down this path to make sure we are not doing more than needed.

Patch-set: 8
Reviewer: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1