Update patch set 2
Patch Set 2:

From what I can tell, it seems to be correct to add the runtime APEX to search paths because it is not a sufficient but rather a necessary condition for a library to be exported for it to appear in search paths. In order for a library to be exported from a namespace, it also needs to be present in shared_libs.

Patch-set: 2
Reviewer: Gerrit User 1067098 <1067098@85c56323-6fa9-3386-8a01-6480fb634889>