Update patch set 3
Patch Set 3:

> Patch Set 3: Presubmit-Ready+1
> 
> Dropped destroying the dm node, skip resize if node instantiated (and thus likely in use)

I realized late yesterday that PS2 was safe. We were worried about overwriting blocks on an in-use filesystem. But deleting the node will fail if the filesystem is mounted, so there's no risk. In fact it's safer to do this, since we create *all* partitions in the first-stage, not just the current slot.

Patch-set: 3
Reviewer: Gerrit User 1290458 <1290458@85c56323-6fa9-3386-8a01-6480fb634889>