Update patch set 6
Patch Set 6:

I took a look at cuttlefish today, and the fstab does indeed reference partitions outside of what it provides for boot_devices.  It has a separate 'device' for each of its partitions and it uses boot_devices to indicate where the super partition is, since that's what we required for dm-linear.  It has its metadata partitions, required in first stage mount, in a different device.  As a side note, this does mean that /dev/block/by-name is a bit broken on CF.

If we want to move forward with this, then we need a way around that.

Alternatively, I'm still interested in knowing if there's something else that Intel can do to work around this?  If DT is used (or something equivalent), could the bootloader 'disable' the nodes of the devices that are not planned to be used?  We've had first stage mount for a few releases now and never explicitly supported the idea that there would be multiple block devices with the same partition names, and honestly, I think it may be best to not support that use case.

Patch-set: 6
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>