Update patch set 1
Patch Set 1:

> I hope that there is also support for non-errno cases.

We can always rename errno to 'code' and have it be generic, but keep ErrnoError as it is.

> Are you really happy with c_str() on every string arguments? :(

Not really; though I think the biggest gain with streams is automatically handling integer types correctly and the ability to write complex objects that have their own operator<<.

> I am also worried about "a" + "b" + "c" because of its performance.

Streams aren't super efficient either, but as you said, it's the error case, so probably not something that needs much optimization.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>