Update patch set 5
Patch Set 5:

I'm trying not to necessarily insist on MS_MOVE; if you really do think the two bind mounts are better, then lets go with that.  You have more experience with these CLs than I do.

> a voice in my head worries about open file descriptors being invalidated by the umounts

MNT_DETACH ensures that open FDs will stay valid.  There's only a concern if, for example, someone has a dirfd and is trying to use openat(), but that's hopefully a rare case, as it is certainly broken by the bind mount use case.

> what moves to ignore and/or the order that operations are performed in.

Note that bind mounts do require strict ordering.  Restoring the bind mounts needs to happen from least specific to most specific, otherwise the mounts will be shadowed just like the overlayfs case.  As long as we're iterating in a forwards direction and assume /proc/mounts returns the mounts in this order, we're okay, but it does require that direction and that assumption.  It's actually MS_MOVE, once we prune the list, that allows complete freedom from ordering.

It really is unfortunate that the 'shared' bit isn't provided in /proc/mounts.  

The marking private / shared part I think is indeed complex.  I'm just thinking about the code in PS2: we only marked the mount point as shared again after we did the overlayfs mount, right?  I'm not quite sure that is right, since after that point when referring to 'mount_point', we're actually referring to the overlayfs mount.  Though, since the original /vendor mount is hidden by the overlayfs mount, it's not probably not important to mark it as shared anymore (since no one can mount under it).

That said, is the overlayfs mount created as shared?  It probably should be, and we should have it as 'shared' while moving (or binding) the submounts back into it.  If we don't, then other mount namespaces won't see these submounts being moved (binded) back into it.  We can move mounts as children into a shared mount parent, just not move them out.

Patch-set: 5
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>