Update patch set 5
Patch Set 5:

We can let this one percolate, we all have to be happy.

[TL;DR]

This is a compact/KISS fix adb remount, but I am _not_ 100% convinced it is needed with any urgency as it solves a corner case no one has yet complained about (except me). That is what I feel most comfortable (currently) for BC if the corner cases it covers are viewed serious enough. This fixes two corner cases: eng builds that require phone services (firmware_mnt) and device configurations that mount sub-mounts in first stage init; neither are a usage pattern in the bug, although they are related to the bug. The other CL already fixes all the more common use cases.

For the future, re-evaluate and get the shared status, consider switching back to MS_MOVE because it has the most likely chance to preserve most open references; or any other mechanism that can demonstrably deal with the race conditions that I currently do not have an explanation/fix for. Without this change, things are worse than the race condition (the related sub-mount content disappears and does not come back).

The real(tm) fix may be the overlayfs driver.

I believe the overlayfs mount is 'shared' because that is typically the automatic default unless inside a container. Another assumption. Neither first stage init or adb are inside a container. The worry here is that the sub-mount could be _private_ (?) or someone runs the remount command in a container (?).

Plan: (with less pressure)
- Investigate parsing /proc/self/mountinfo instead of /proc/mounts
- Investigate MS_MOVE, FS_SHARED, FS_PRIVATE and what they really mean in the context of Android. (admit mount namespaces and containers

Patch-set: 5
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>