Update patch set 14
Patch Set 14:

FWIW, I'd recommend removing this change from the topic and submitting it separately now, since it seems self contained and won't be enabled unless the right property is set, so it should be a no-op.  And if it isn't a no-op, then better to know that sooner than later.

You may need to break out the rootdir/ parts of the change, but that's trivial.

Patch-set: 14
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>