Update patch set 4
Patch Set 4:

This code was not done because of a security boundary, that is but a carefully crafted side effect. packagelist_parse at one point was not thread safe (and I had no interest in fixing it at the time and the cleanup bug I think was punted). Please confirm that it is thread safe now.

Patch-set: 4
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>