Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> can they be statically linked then?

Vendor code can statically link them, but we don't want them to do this. We want all ashmem FDs to be obtained and manipulated via the libcutils API. Once  this happens we can begin replacing /dev/ashmem with memfd. So we break vendors, and clearly communicate to them that the fix is to use to libcutils instead of opening /dev/ashmem directly.

Patch-set: 3
Reviewer: Gerrit User 1499410 <1499410@85c56323-6fa9-3386-8a01-6480fb634889>