Update patch set 3
Patch Set 3:

(1 comment)

Do we want to do this?  It's only a bit shorter than the << operator, but is less intuitive.  The stream operator parallels LOG() <<, so at least people would be used to it.

How would this interact with `return Error("%s %d %s", a, 12, c);" ?  Assuming that it may be harder to differentiate, I'd recommend dropping this and just giving 2 options: streams as is, and passing all of the arguments of Error() directly to StringPrintf (or, even better, fmtlib, or something similar).

Patch-set: 3
CC: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>