Update patch set 5
Patch Set 5:

(1 comment)

(I like the change, not clear how caller is going to determine if its not a raw device. May be we should also add a IsDeviceMapperDevice() in libdm that doesn't depend on the name prefix, but actually checks with the kernel.

Patch-set: 5
Reviewer: Gerrit User 1108565 <1108565@85c56323-6fa9-3386-8a01-6480fb634889>