Update patch set 1
Patch Set 1:

I think that we need to go back to either having a sequence number or, probably better, tracking the monotonic timestamp of each entry.  Just relying on the realtime time stamp is not enough, since there are plenty of bad cases we can get into.

I'm just not sure if it's better to track the monotonic timestamp of when the entry reaches logd, which would be a slightly better sequence number, or if we should take a timestamp supplied by the client, which would be more accurate, but could be manipulated by clients to insert log messages into arbitrary positions in the log, causing a bunch of potential issues.  For the safety reason, I think the first is the best.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0