Update patch set 2
Patch Set 2:

I would have preferred that fs_mgr_mount_all returns to the main init loop while the process is running in the background, with some sort of async response handler triggered when the process ends - ie generalizing the mechanism that handles launching processes and waiting for properties, as we discussed a couple of years ago. However, I don't have time to implement it so I'm not going to complain about this fix!

Patch-set: 2
Reviewer: Gerrit User 1060907 <1060907@85c56323-6fa9-3386-8a01-6480fb634889>